[PORT FROM R44D] libcamera2: rework EXIF buffer allocations
Previously the EXIF buffer allocation size was based on static
constant of 64KB. But we need to make this dynamic, as we cannot
know the sizes of the additional EXIF APPn marker sections beforehand.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-743
Reviewed-on: https://android.intel.com:443/256860

+++ exif/Exif.h
@@ -323,7 +323,7 @@ typedef struct {
     uint16_t resolution_unit;
     uint16_t compression_scheme;
 
-    uint16_t makerNoteDataSize;
+    uint32_t makerNoteDataSize;
     unsigned char *makerNoteData;
 } exif_attribute_t;
